Move User initial Update Around
